+++ jekyll-gradle-plugin/src/main/kotlin/config/extension.kt
@@ -8,6 +8,7 @@ import dev.petuska.jekyll.util.ProjectEnhancer
 internal fun ProjectEnhancer.configure(extension: JekyllExtension) {
   with(extension) {
     version.convention("latest")
+    environment.convention(System.getenv())
     val sourceSets: JekyllSourceSets = objects.domainObjectContainer(JekyllSourceSet::class.java) { name ->
       objects.newInstance(JekyllSourceSet::class.java, name, objects)
     }