+++ jekyll-gradle-plugin/src/main/kotlin/extension/domain/JekyllSourceSet.kt
@@ -7,9 +7,17 @@ import dev.petuska.jekyll.config.execTaskName
 import dev.petuska.jekyll.config.initTaskName
 import dev.petuska.jekyll.config.resourcesTaskName
 import dev.petuska.jekyll.config.serveTaskName
+import dev.petuska.jekyll.task.BundleExecTask
+import dev.petuska.jekyll.task.JekyllBuildTask
+import dev.petuska.jekyll.task.JekyllExecTask
+import dev.petuska.jekyll.task.JekyllInitTask
+import dev.petuska.jekyll.task.JekyllServeTask
 import org.gradle.api.NamedDomainObjectContainer
 import org.gradle.api.file.SourceDirectorySet
 import org.gradle.api.model.ObjectFactory
+import org.gradle.api.provider.Property
+import org.gradle.api.tasks.Copy
+import org.gradle.language.jvm.tasks.ProcessResources
 import javax.inject.Inject
 
 @Suppress("MemberVisibilityCanBePrivate")
@@ -20,12 +28,19 @@ public abstract class JekyllSourceSet @Inject constructor(public val name: Strin
   public val allSources: SourceDirectorySet = objectFactory.sourceDirectorySet("allsource", "$name sources")
 
   public val processResourcesTaskName: String = resourcesTaskName(name)
+  public abstract val processResourcesTask: Property<ProcessResources>
   public val assembleTaskName: String = assembleTaskName(name)
+  public abstract val assembleTask: Property<Copy>
   public val buildTaskName: String = buildTaskName(name)
+  public abstract val buildTask: Property<JekyllBuildTask>
   public val initTaskName: String = initTaskName(name)
+  public abstract val initTask: Property<JekyllInitTask>
   public val execTaskName: String = execTaskName(name)
+  public abstract val execTask: Property<JekyllExecTask>
   public val bundleExecTaskName: String = bundleExecTaskName(name)
+  public abstract val bundleExecTask: Property<BundleExecTask>
   public val serveTaskName: String = serveTaskName(name)
+  public abstract val serveTask: Property<JekyllServeTask>
 
   init {
     jekyllSourcePatterns.forEach(jekyll.filter::include)
@@ -44,6 +59,9 @@ public abstract class JekyllSourceSet @Inject constructor(public val name: Strin
       "**/*.MARKDOWN",
       "**/*.html",
       "**/*.HTML",
+      "**/*.liquid",
+      "**/*.LIQUID",
+      "**/*.js",
     )
   }
 }