+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# 0.0.2
+
+## Build Versions
+
+* Gradle: 7.4.2
+* JDK: 11
+
+## Changes
+
+* Restructured soureSet layout to better isolate jekyll sources from config resources
+* Intermediate assemble and processResources tasks introduced
+* Improved caching and lazy configuration
+* Expanded documentation and KDoc
+
 # 0.0.1
 
 ## Build Versions